Fixes 5332: Fixed warning on delete completely post. (#5399)
* Fixed warning on delete completely post.

* Added fix for second method

+++ inc/Engine/Preload/Database/Queries/Cache.php
@@ -254,7 +254,9 @@ class Cache extends Query {
 
 		$deleted = true;
 		foreach ( $items as $item ) {
-			$deleted = $deleted && $this->delete_item( $item->id );
+			if ( ! is_bool( $item ) ) {
+				$deleted = $deleted && $this->delete_item( $item->id );
+			}
 		}
 
 		return $deleted;
@@ -290,7 +292,9 @@ class Cache extends Query {
 		$rows = $this->get_old_cache();
 
 		foreach ( $rows as $row ) {
-			$this->delete_item( $row->id );
+			if ( ! is_bool( $row ) ) {
+				$this->delete_item( $row->id );
+			}
 		}
 	}
